my-notepad-app/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
Flask_SQLAlchemy==3.1.1
email_validator==2.2.0
python-slugify==8.0.4
